Refactor: Alpha의 value 프로퍼티의 타입 Double -> Float 변경
ViewController의 alphaSliderValueChanged(_:) 액션 메서드에서 변경될 alpha값을 변환하는 과정을 줄이기 위해 Alpha의 value 프로퍼티 타입을 Float로 변경
그로 인한 다른 메서드들의 매개변수 타입 수정
Plane 구조체의 stepUpAlphaValue(), stepDownAlphaValue() 메서드에서 사용하는 newAlpha 상수의 값을 결정할때 Float 타입 값들끼리 뺄셈을 함으로 인해 소수의 값에 필요없는 값이 추가되어 해당 값 처리하는 과정을 추가

+++ DrawingApp/DrawingApp/Model/ViewFactory.swift
@@ -24,11 +24,11 @@ class ViewFactory {
 }
 
 extension BackgroundColor {
-    fileprivate func convertToUIColor(with alphaValue: Double = 1.0) -> UIColor {
+    fileprivate func convertToUIColor(with alphaValue: Float = 1.0) -> UIColor {
         let convertedRed = Double(self.r) / 255.0
         let convertedGreen = Double(self.g) / 255.0
         let convertedBlue = Double(self.b) / 255.0
         
-        return UIColor(red: convertedRed, green: convertedGreen, blue: convertedBlue, alpha: alphaValue)
+        return UIColor(red: convertedRed, green: convertedGreen, blue: convertedBlue, alpha: CGFloat(alphaValue))
     }
 }
